requirements.txt
llama-index
python-dotenv
paho-mqtt